// types..d.ts
export type ServiceTypeCustom = 'node' | 'rust' | 'python' | 'golang' | 'php';

export type ServiceTypeCommon =
  | 'redis'
  | 'postgres'
  | 'mysql'
  | 'adminer'
  | 'mariadb'
  | 'mongo'
  | 'rabbitmq'
  | 'phpmyadmin'
  | 'pgadmin'
  | 'mongo_express';

export type ServiceTypeCommonPublic = 'adminer' | 'phpmyadmin' | 'pgadmin' | 'mongo_express';

export type ServiceType = ServiceTypeCommon | ServiceTypeCustom;

export type ServiceSize =
  | 'pico'
  | 'nano'
  | 'micro'
  | 'mili'
  | 'santi'
  | 'deci'
  | 'deca'
  | 'hecto'
  | 'kilo';

export type PortType = 'http' | 'ws' | 'chunked' | 'php';

export type GitUntrackedPolicy = 'checkout' | 'push' | 'merge';

export type Domains = Record<string, string>;

export interface Port {
  port: number;
  type: PortType;
  location?: string;
  timeout?: string;
  buffer_size?: string;
  proxy_path?: string;
  static?: {
    location: string;
    path: string;
    index?: string;
  }[];
}
// ...
export type GitType = 'github' | 'gitlab';